Rename and export getKeysets
